refactor: Extract functions out of markdown#create
The whole creation of the markdown file was happening in the body of the
create method. This is an attempt to extract the distinct sections that
every changelog will have, so now there is a function for each of the
sections:

* The title
* Story summary
* Commits

Also, having those as functions might provide easier and more clear ways
to handle them individually (i.e. suppoting options)