feat: Introduce the Config struct
The idea will be for this struct to hold the default and supported tags
and future configuration (e.g. default output filename, tracker).

As part of this I also want to group the tags by their description, and
for that to happen the current approach of having one regex with all
tag patterns on it might need to change.