feat: Use new tagging
Commits, Changelog and Formatter will now leverage the named tags
feature. Tags with the same name will be grouped together, regardless of
the number of patterns associated to it.

Story numbers will be considered the component of such tag therefore
those need to be captured as a group in the regex.